requirements/pip.txt
django
django-filter
djangorestframework
nodeenv
